- moved contents of temba/msgs/attachmentscompose.py to temba/utils/fields.py
